Standardize groupId to match Maven convention.
This change will help enable eventual distribution of ImageJ through the
Maven central repository. The requirement for groupId is that it
correspond to Java packaging convention of reverse domain name. Since we
registered imagej.net and plan for that domain to serve as the central
ImageJ resource, the groupId needs to be net.imagej.

In the future, before the final release of ImageJ 2.0.0, we should also
consider whether to use net.imagej as the Java package prefix as well.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4234 be23795f-a60e-40d4-ba19-055798872e57