Fix Roi setIntBounds widths
When converting from double to int bounds, for the starting point we
take the floor of the double starting point + 0.5. This is because rois
are drawn "around" the integer bounding pixels, essentially "on" the
half-pixel space. If a bound's top-left corner is (10.5, 10.5) then the
corresponding integer bound corner is (11, 11), as the corresponding
rectangle is drawn "through" (10.5, 10.5) to surround the pixel at (11,
11).

For the int bound's end point we should take the floor of the double end
point + 1. A double bound end point of (20, 20) goes through the middle
of the (20, 20) pixel, so our int bound needs to extend to (21, 21) to
be drawn "through" (20.5, 20.5) and contain the double bound end point.
A double bound end point of (19.5, 19.5) is rounded up to (20, 20) with
this logic, which is appropriate.

This fixes a bug where we were adding 1.5 to the double bound end
points, which resulted in an int bound one pixel too large for the cases
the double bound ended in a decimal >= 5.