Update ImagesToStack.java to retain all overlays when not all images are the same size
Previously it was not possible to combine multiple images of different sizes and overlays into a stack with this function.
All but the overlay of the first image was lost when doing this.
Now intuitive behaviour is implemented to retain all overlays when using the copy method (center or top-left).
Application: e.g. labelled cells (via overlay) in a temporal stack of stitched images that slightly vary in image dimensions over time.